make testZeekHandler available without zeek build flag
