+++ zqd/handlers_test.go
@@ -5,6 +5,7 @@ import (
 	"context"
 	"encoding/json"
 	"errors"
+	"io"
 	"io/ioutil"
 	"net/http/httptest"
 	"os"
@@ -20,6 +21,7 @@ import (
 	"github.com/brimsec/zq/zio/zngio"
 	"github.com/brimsec/zq/zqd"
 	"github.com/brimsec/zq/zqd/api"
+	"github.com/brimsec/zq/zqd/zeek"
 	"github.com/brimsec/zq/zql"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
@@ -457,4 +459,65 @@ func postSpaceLogs(t *testing.T, c *api.Connection, spaceName string, logs ...st
 		payloads = append(payloads, p)
 	}
 	return payloads
+}
+
+func testZeekLauncher(start, wait procFn) zeek.Launcher {
+	return func(ctx context.Context, r io.Reader, dir string) (zeek.Process, error) {
+		p := &testZeekProcess{
+			ctx:    ctx,
+			reader: r,
+			wd:     dir,
+			wait:   wait,
+			start:  start,
+		}
+		return p, p.Start()
+	}
+}
+
+type procFn func(t *testZeekProcess) error
+
+type testZeekProcess struct {
+	ctx    context.Context
+	reader io.Reader
+	wd     string
+	start  procFn
+	wait   procFn
+}
+
+func (p *testZeekProcess) Start() error {
+	if p.start != nil {
+		return p.start(p)
+	}
+	return nil
+}
+
+func (p *testZeekProcess) Wait() error {
+	if p.wait != nil {
+		return p.wait(p)
+	}
+	return nil
+}
+
+func writeLogsFn(logs []string) procFn {
+	return func(t *testZeekProcess) error {
+		for _, log := range logs {
+			r, err := os.Open(log)
+			if err != nil {
+				return err
+			}
+			defer r.Close()
+			base := filepath.Base(r.Name())
+			w, err := os.Create(filepath.Join(t.wd, base))
+			if err != nil {
+				return err
+			}
+			defer w.Close()
+			if _, err = io.Copy(w, r); err != nil {
+				return err
+			}
+		}
+		// drain the reader
+		_, err := io.Copy(ioutil.Discard, t.reader)
+		return err
+	}
 }
